ignores node modules

+++ models/floatrequest.go
@@ -3,8 +3,9 @@ package models
 import "time"
 
 type FloatRequest struct {
-	ID        uint `gorm:"primaryKey"`
-	Amount    float64
+	ID uint `gorm:"primaryKey"`
+	// Amount    float64
+	Amount    int
 	Status    string
 	CreatedAt time.Time
 }
